+++ src/snowflake/cli/_plugins/sql/commands.py
@@ -14,6 +14,7 @@
 
 from __future__ import annotations
 
+import sys
 from pathlib import Path
 from typing import List, Optional
 
@@ -28,6 +29,7 @@ from snowflake.cli.api.commands.snow_typer import SnowTyperFactory
 from snowflake.cli.api.commands.utils import parse_key_value_variables
 from snowflake.cli.api.output.types import (
     CommandResult,
+    MessageResult,
     MultipleResults,
     QueryResult,
 )
@@ -42,7 +44,7 @@ SourceOption = OverrideableOption(
 )
 
 
-@app.command(name="sql", requires_connection=True, no_args_is_help=True)
+@app.command(name="sql", requires_connection=True, no_args_is_help=False)
 @with_project_definition(is_optional=True)
 def execute_sql(
     query: Optional[str] = SourceOption(
@@ -90,20 +92,33 @@ def execute_sql(
     if data_override:
         data = {v.key: v.value for v in parse_key_value_variables(data_override)}
 
+    retain_comments = bool(retain_comments)
+    std_in = bool(std_in)
+
+    no_source_provided = not any([query, files, std_in])
+    if no_source_provided and not sys.stdin.isatty():
+        maybe_pipe = sys.stdin.read().strip()
+        if maybe_pipe:
+            query = maybe_pipe
+            std_in = True
+
+    if no_source_provided:
+        from snowflake.cli._plugins.sql.repl import Repl
+
+        Repl(SqlManager(), data=data, retain_comments=retain_comments).run()
+        return MessageResult("")
+
     expected_results_cnt, cursors = SqlManager().execute(
         query, files, std_in, data=data, retain_comments=retain_comments
     )
     if expected_results_cnt == 0:
         # case expected if input only scheduled async queries
-        # ends gracefully with no message for consistency with snowsql.
-        import sys
-
         list(cursors)  # evaluate the result to schedule potential async queries
+        # ends gracefully with no message for consistency with snowsql.
         sys.exit(0)
 
     if expected_results_cnt == 1:
         # evaluate the result to schedule async queries
         return QueryResult(list(cursors)[0])
 
-    # TODO: ckeck format json
     return MultipleResults((QueryResult(c) for c in cursors))