+++ src/snowflake/cli/_plugins/sql/manager.py
@@ -27,6 +27,7 @@ from snowflake.cli._plugins.sql.source_reader import (
     files_reader,
     query_reader,
 )
+from snowflake.cli.api.cli_global_context import get_cli_context
 from snowflake.cli.api.console import cli_console
 from snowflake.cli.api.exceptions import CliArgumentError, CliSqlError
 from snowflake.cli.api.rendering.sql_templates import snowflake_sql_jinja_render
@@ -84,9 +85,10 @@ class SqlManager(SqlExecutionMixin):
                 cli_console.warning(error)
             raise CliSqlError("SQL rendering error")
 
+        cursor_class = SnowflakeCursor if get_cli_context().is_repl else VerboseCursor
         return expected_results_cnt, self._execute_compiled_statements(
             compiled_statements,
-            cursor_class=VerboseCursor,
+            cursor_class=cursor_class,
         )
 
     def _execute_compiled_statements(